Moved panel related stuff to panel_utils file
