+++ core/animations/test_panels.py
@@ -4,117 +4,8 @@ from scipy import signal
 import numpy as np
 import time
 import random
-
 import colorsys
-
-
-''' right 8 extra pixel '''
-
-fmap = {
-    'stats' : {
-        'total_pixels' : 1025,
-        'r_pixels' : 319,
-        'c_pixels' : 386,
-        'l_pixels' : 320
-    },
-    'right' : [[0,   27],
-                [28,  43],
-                [44,  65],
-                [66,  81],
-                [82,  91],
-                [92,  111],
-                [112, 135],
-                [136, 152],
-                [153, 165],
-                [166, 191],
-                [192, 214],
-                [215, 240],
-                [241, 263],
-                [264, 298],
-                [299, 318]],
-    'centre' : [[0,   27],
-                [28,  53],
-                [54,  66],
-                [67,  80],
-                [81,  89],
-                [90,  102],
-                [103, 128],
-                [129, 155],
-                [156, 224],
-                [225, 241],
-                [242, 272],
-                [273, 289],
-                [290, 318],
-                [319, 338],
-                [339, 355],
-                [356, 385]],
-    'left'  :   [[0,   27],
-                [28,  43],
-                [44,  65],
-                [66,  81],
-                [82,  91],
-                [92,  111],
-                [112, 136],
-                [137, 153],
-                [154, 166],
-                [167, 192],
-                [193, 215],
-                [216, 241],
-                [242, 264],
-                [265, 299],
-                [300, 319]]
-}
-
-
-
-class FaceSection():
-
-
-
-    def __init__(self, length=10,section='left'):
-
-
-        self.length = length
-        self.section = section
-
-
-        self.cycle_start = time.time()
-
-        self.temp_pixels = np.array([[0.0,0.0,0.0]]*self.length)
-
-        self.logger = logging_handler_setup('face section - %s'%self.section)
-
-        
-
-    def update(self, fft, carrier):
-
-        if self.check_beat(fft[8:10]):
-            panel_it = random.randint(0,len(fmap[self.section])-1)
-            h = random.uniform(carrier-.1, carrier+.1)
-
-            for x in xrange(fmap[self.section][panel_it][0],fmap[self.section][panel_it][1]+1):
-                self.temp_pixels[x] = colorsys.hsv_to_rgb(h,.9,.9)
-
-        self.temp_pixels = self.decay_pixels()
-        return self.temp_pixels
-
-        
-    def decay_pixels(self):
-        i = np.vectorize(self.add_random)
-        return i(self.temp_pixels,-.01,0.0)
-
-        
-    def add_random(self,x,lower,upper):
-        return x + random.uniform(lower,upper)
-
-
-    def check_beat(self, beats):
-        if sum(beats) > 0:
-            return True
-        else:
-            return False
-
- 
+from panel_utils import fmap, FaceSection
 
 class TestPanels(Animation):
     layout_type = "Layout"
